# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Influencer Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /influencers:
    get:
      operationId: listInfluencers
      tags: [influencers]
      summary: List influencers with optional filters
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: minFollowers
          in: query
          schema:
            type: number
        - name: maxFollowers
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of influencers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Influencer"
  /influencers/{id}:
    get:
      operationId: getInfluencer
      tags: [influencers]
      summary: Get influencer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Influencer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InfluencerDetail"
        "404":
          description: Not found
    patch:
      operationId: updateInfluencer
      tags: [influencers]
      summary: Update influencer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InfluencerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Influencer"

  # ─── Brands ──────────────────────────────────────────────────────────────────
...
  /conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List conversations for current user
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [messages]
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ─── Notifications ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Influencer:
      type: object
      required: [id, userId, name, category, country, followers, engagementRate, avgViews, collaborationCost, platforms, languages, avatarUrl]
      properties:
        id:
          type: number
        userId:
          type: number
        name:
          type: string
        bio:
          type: ["string", "null"]
        category:
          type: string
        country:
          type: string
        followers:
          type: number
        engagementRate:
          type: number
        avgViews:
          type: number
        collaborationCost:
          type: number
        platforms:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        coverUrl:
          type: ["string", "null"]
        profileCompletion:
          type: number
        monthlyEarnings:
          type: number
        isVerified:
          type: boolean
    InfluencerDetail:
      allOf:
        - $ref: "#/components/schemas/Influencer"
        - type: object
          properties:
            portfolio:
              type: array
              items:
                type: string
            previousCollaborations:
              type: array
              items:
                $ref: "#/components/schemas/Collaboration"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            availability:
              type: string
    Collaboration:
      type: object
      required: [id, brandName, campaignTitle, year]
      properties:
        id:
          type: number
        brandName:
          type: string
        campaignTitle:
          type: string
        year:
          type: number
    Review:
      type: object
      required: [id, brandName, rating, comment, createdAt]
      properties:
        id:
          type: number
        brandName:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
    InfluencerUpdate:
      type: object
      properties:
        bio:
          type: string
        category:
          type: string
        country:
          type: string
        collaborationCost:
          type: number
        platforms:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        coverUrl:
          type: string
        availability:
          type: string

    # ─── Brand ────────────────────────────────────────────────────────────────
...
    Conversation:
      type: object
      required: [id, participantId, participantName, participantAvatarUrl, lastMessage, lastMessageAt, unreadCount]
      properties:
        id:
          type: number
        participantId:
          type: number
        participantName:
          type: string
        participantAvatarUrl:
          type: ["string", "null"]
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: number
    ConversationInput:
      type: object
      required: [participantId]
      properties:
        participantId:
          type: number
    Message:
      type: object
      required: [id, conversationId, senderId, content, createdAt]
      properties:
        id:
          type: number
        conversationId:
          type: number
        senderId:
          type: number
        senderName:
          type: ["string", "null"]
        senderAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ─── Notification ─────────────────────────────────────────────────────────
...