# lib/api-zod/src/generated/types/collaboration.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Influencer Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Collaboration {
  id: number;
  brandName: string;
  campaignTitle: string;
  year: number;
}