# lib/api-zod/src/generated/types/influencerDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Influencer Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Collaboration } from './collaboration';
import type { Influencer } from './influencer';
import type { Review } from './review';

export type InfluencerDetail = Influencer & {
  portfolio?: string[];
  previousCollaborations?: Collaboration[];
  reviews?: Review[];
  availability?: string;
};